hashbreaker.py
import re
import urllib3
import os
import requests
import concurrent.futures
import websocket

urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

# Colors for CLI (optional)
end = '\033[0m'
red = '\033[91m'
green = '\033[92m'
white = '\033[97m'
dgreen = '\033[32m'
yellow = '\033[93m'
back = '\033[7;91m'
run = '\033[97m[~]\033[0m'
que = '\033[94m[?]\033[0m'
bad = '\033[91m[-]\033[0m'
info = '\033[93m[!]\033[0m'
good = '\033[92m[+]\033[0m'

# --- API Functions ---
def alpha(hashvalue, hashtype):
    # your existing alpha function code here
    pass

def beta(hashvalue, hashtype):
    # your existing beta function code here
    pass

def gamma(hashvalue, hashtype):
    # your existing gamma function code here
    pass

def theta(hashvalue, hashtype):
    # your existing theta function code here
    pass

# --- Main crack function ---
def crack(hashvalue):
    result = False
    md5 = [alpha,beta,gamma,theta]
    sha1 =[alpha,beta,theta]
    sha256 = [alpha, beta, theta]
    sha384 = [alpha, beta, theta]
    sha512 = [alpha, beta, theta]

    if len(hashvalue) == 32:
        for api in md5:
            r = api(hashvalue, 'md5')
            if r:
                return r
    elif len(hashvalue) == 40:
        for api in sha1:
            r = api(hashvalue, 'sha1')
            if r:
                return r
    elif len(hashvalue) == 64:
        for api in sha256:
            r = api(hashvalue, 'sha256')
            if r:
                return r
    elif len(hashvalue) == 96:
        for api in sha384:
            r = api(hashvalue, 'sha384')
            if r:
                return r
    elif len(hashvalue) == 128:
        for api in sha512:
            r = api(hashvalue, 'sha512')
            if r:
                return r
    else:
        return False

# --- CLI Execution ---
if __name__ == "__main__":
    import argparse

    parser = argparse.ArgumentParser(description="HashBreaker CLI")
    parser.add_argument('-s', help='hash', dest='hash')
    parser.add_argument('-f', help='file containing hashes', dest='file')
    parser.add_argument('-d', help='directory containing hashes', dest='dir')
    parser.add_argument('-t', help='number of threads', dest='threads', type=int)
    args = parser.parse_args()

    # Your existing CLI code goes here, e.g., reading file/directory or single hash
