requirements.txt
Flask==2.3.3
requests>=2.28.1
urllib3>=1.26.12
websocket-client>=1.3.3
gunicorn==21.2.0
